+++ apps/playful-micro/src/App.tsx
@@ -1,0 +1,7 @@
+import { PlayfulMicroDemo } from "./components/PlayfulMicroDemo";
+
+function App() {
+  return <PlayfulMicroDemo />;
+}
+
+export default App;